Unique Id's should now actually be Unique

+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/statistics/ui/homecards/cards/StatisticsCardItem.kt
@@ -31,6 +31,6 @@ data class LocalStatisticsCardItem(
     val onRemoveListener: (LocalIncidenceStats) -> Unit,
 ) : StatisticsCardItem() {
     override val stableId: Long = when (stats) {
-        is LocalIncidenceStats -> stats.selectedLocation.uniqueID.toLong()
+        is LocalIncidenceStats -> stats.selectedLocation.uniqueID
     }
 }
